+++ agent/engine/docker_task_engine.go
@@ -46,6 +46,7 @@ import (
 	"github.com/aws/amazon-ecs-agent/agent/taskresource/credentialspec"
 	"github.com/aws/amazon-ecs-agent/agent/taskresource/firelens"
 	"github.com/aws/amazon-ecs-agent/agent/utils"
+	referenceutil "github.com/aws/amazon-ecs-agent/agent/utils/reference"
 	"github.com/aws/amazon-ecs-agent/ecs-agent/api/appnet"
 	apicontainerstatus "github.com/aws/amazon-ecs-agent/ecs-agent/api/container/status"
 	apierrors "github.com/aws/amazon-ecs-agent/ecs-agent/api/errors"
@@ -60,6 +61,7 @@ import (
 	ep "github.com/aws/aws-sdk-go/aws/endpoints"
 	"github.com/docker/docker/api/types"
 	dockercontainer "github.com/docker/docker/api/types/container"
+	"github.com/opencontainers/go-digest"
 	"github.com/pkg/errors"
 )
 
@@ -81,6 +83,7 @@ const (
 	ipamCleanupTmeout                  = 5 * time.Second
 	minEngineConnectRetryDelay         = 2 * time.Second
 	maxEngineConnectRetryDelay         = 200 * time.Second
+	tagImageTimeout                    = 30 * time.Second
 	engineConnectRetryJitterMultiplier = 0.20
 	engineConnectRetryDelayMultiplier  = 1.5
 	// logDriverTypeFirelens is the log driver type for containers that want to use the firelens container to send logs.
@@ -979,7 +982,7 @@ func (engine *DockerTaskEngine) EmitTaskEvent(task *apitask.Task, reason string)
 	event, err := api.NewTaskStateChangeEvent(task, reason)
 	if err != nil {
 		if _, ok := err.(api.ErrShouldNotSendEvent); ok {
-			logger.Debug(err.Error())
+			logger.Debug(err.Error(), logger.Fields{field.TaskID: task.GetID()})
 		} else {
 			logger.Error("Unable to create task state change event", logger.Fields{
 				field.TaskID: task.GetID(),
@@ -1239,12 +1242,134 @@ func (engine *DockerTaskEngine) GetDaemonManagers() map[string]dm.DaemonManager 
 func (engine *DockerTaskEngine) pullContainerManifest(
 	task *apitask.Task, container *apicontainer.Container,
 ) dockerapi.DockerContainerMetadata {
-	// Currently a no-op
-	logger.Debug("Manifest pull is currently a no-op", logger.Fields{
-		field.TaskID:    task.GetID(),
-		field.Container: container.Name,
-		field.Image:     container.Image,
+	if !container.DigestResolutionRequired() {
+		// Digest resolution not required
+		// (internal container or already has digest in image reference)
+		logger.Info("Digest resolution not required", logger.Fields{
+			field.TaskARN:       task.Arn,
+			field.ContainerName: container.Name,
+			field.Image:         container.Image,
+		})
+		return dockerapi.DockerContainerMetadata{}
+	}
+	// AppNet Agent container image is managed at start up so it is not in-scope for digest resolution.
+	// (it uses the same image as AppNet Relay container)
+	if task.IsServiceConnectEnabled() && container == task.GetServiceConnectContainer() {
+		return dockerapi.DockerContainerMetadata{}
+	}
+
+	var imageManifestDigest digest.Digest
+	if !engine.imagePullRequired(engine.cfg.ImagePullBehavior, container, task.GetID()) {
+		// Image pull is not required for the container so we will use a locally available
+		// image for the container. Get digest from a locally available image.
+		imgInspect, err := engine.client.InspectImage(container.Image)
+		if err != nil {
+			logger.Error("Failed to inspect image to find repo digest", logger.Fields{
+				field.TaskARN:       task.Arn,
+				field.ContainerName: container.Name,
+				field.Image:         container.Image,
+			})
+			return dockerapi.DockerContainerMetadata{
+				Error: dockerapi.CannotPullImageManifestError{FromError: err},
+			}
+		}
+		if len(imgInspect.RepoDigests) == 0 {
+			// Image was not pulled from a registry, so the user must have cached it on the
+			// host directly. Skip digest resolution for this case as there is no digest.
+			logger.Info("No repo digest found", logger.Fields{
+				field.TaskARN:       task.Arn,
+				field.ContainerName: container.Name,
+				field.Image:         container.Image,
+			})
+			return dockerapi.DockerContainerMetadata{}
+		}
+		parsedDigest, err := referenceutil.GetDigestFromRepoDigests(
+			imgInspect.RepoDigests, container.Image)
+		if err != nil {
+			logger.Error("Failed to find a repo digest matching the image", logger.Fields{
+				field.TaskARN:       task.Arn,
+				field.ContainerName: container.Name,
+				field.Image:         container.Image,
+				"repoDigests":       imgInspect.RepoDigests,
+			})
+			return dockerapi.DockerContainerMetadata{
+				Error: dockerapi.CannotPullImageManifestError{
+					FromError: fmt.Errorf("failed to find a repo digest matching '%s'", container.Image),
+				},
+			}
+		}
+		imageManifestDigest = parsedDigest
+		logger.Info("Fetched image manifest digest for container from local image inspect", logger.Fields{
+			field.TaskARN:       task.Arn,
+			field.ContainerName: container.Name,
+			field.ImageDigest:   imageManifestDigest.String(),
+			field.Image:         container.Image,
+		})
+	} else {
+		// Digest should be resolved by calling the registry
+
+		// Technically, API version 1.30 is enough to call Distribution API to fetch image manifests
+		// from registries. However, Docker engine versions between 17.06 (API version 1.30) and
+		// 17.12 (API version 1.35) support image pulls from v1 registries
+		// (using `disable-legacy-registry` option) whereas Distribution API does not work against
+		// v1 registries. So, to be safe, we will only attempt digest resolution if Docker engine
+		// version is >= 17.12 (API version 1.35).
+		supportedAPIVersion := dockerclient.GetSupportedDockerAPIVersion(dockerclient.Version_1_35)
+		client, err := engine.client.WithVersion(supportedAPIVersion)
+		if err != nil {
+			logger.Warn(
+				"Failed to find a supported API version that supports manifest pulls. Skipping digest resolution.",
+				logger.Fields{
+					field.TaskARN:        task.Arn,
+					field.ContainerName:  container.Name,
+					"requiredAPIVersion": supportedAPIVersion,
+					field.Error:          err,
+				})
+			return dockerapi.DockerContainerMetadata{}
+		}
+
+		// Set registry auth credentials if required and clear them when no longer needed
+		clearCreds, authError := engine.setRegistryCredentials(container, task)
+		if authError != nil {
+			logger.Error("Failed to set registry auth credentials", logger.Fields{
+				field.TaskARN:       task.Arn,
+				field.ContainerName: container.Name,
+				field.Error:         authError,
+			})
+			return dockerapi.DockerContainerMetadata{Error: authError}
+		}
+		if clearCreds != nil {
+			defer clearCreds()
+		}
+
+		ctx, cancel := context.WithTimeout(engine.ctx, engine.cfg.ManifestPullTimeout)
+		defer cancel()
+		distInspect, manifestPullErr := client.PullImageManifest(
+			ctx, container.Image, container.RegistryAuthentication)
+		if manifestPullErr != nil {
+			logger.Error("Failed to fetch image manifest from registry", logger.Fields{
+				field.TaskARN:       task.Arn,
+				field.ContainerName: container.Name,
+				field.Image:         container.Image,
+				field.Error:         manifestPullErr,
+			})
+			return dockerapi.DockerContainerMetadata{Error: manifestPullErr}
+		}
+		imageManifestDigest = distInspect.Descriptor.Digest
+		logger.Info("Fetched image manifest digest for container from registry", logger.Fields{
+			field.TaskARN:       task.Arn,
+			field.ContainerName: container.Name,
+			field.ImageDigest:   imageManifestDigest.String(),
+			field.Image:         container.Image,
+		})
+	}
+
+	logger.Debug("Setting image digest on container", logger.Fields{
+		field.TaskARN:       task.Arn,
+		field.ContainerName: container.Name,
+		field.ImageDigest:   imageManifestDigest.String(),
 	})
+	container.SetImageDigest(imageManifestDigest.String())
 	return dockerapi.DockerContainerMetadata{}
 }
 
@@ -1313,7 +1438,7 @@ func (engine *DockerTaskEngine) imagePullRequired(imagePullBehavior config.Image
 		// by inspecting the image.
 		_, err := engine.client.InspectImage(container.Image)
 		if err != nil {
-			logger.Info("Image inspect returned error, going to pull image for container", logger.Fields{
+			logger.Info("Image inspect returned error, manifest and image pull required", logger.Fields{
 				field.TaskID:    taskId,
 				field.Container: container.Name,
 				field.Image:     container.Image,
@@ -1398,53 +1523,88 @@ func (engine *DockerTaskEngine) pullAndUpdateContainerReference(task *apitask.Ta
 		return dockerapi.DockerContainerMetadata{Error: TaskStoppedBeforePullBeginError{task.Arn}}
 	}
 
-	// Set the credentials for pull from ECR if necessary
-	if container.ShouldPullWithExecutionRole() {
-		executionCredentials, ok := engine.credentialsManager.GetTaskCredentials(task.GetExecutionCredentialsID())
-		if !ok {
-			logger.Error("Unable to acquire ECR credentials to pull image for container", logger.Fields{
-				field.TaskID:    task.GetID(),
-				field.Container: container.Name,
-				field.Image:     container.Image,
+	// Set registry auth credentials if required and clearCreds them when no longer needed
+	clearCreds, err := engine.setRegistryCredentials(container, task)
+	if err != nil {
+		return dockerapi.DockerContainerMetadata{Error: err}
+	}
+	if clearCreds != nil {
+		defer clearCreds()
+	}
+
+	imageRef := container.Image
+	imageDigest := container.GetImageDigest()
+	if imageDigest != "" {
+		// If image digest is available then prepare a canonical reference to be used for image pull.
+		// This ensures that the image version referenced by the digest is pulled.
+		canonicalRef, err := referenceutil.GetCanonicalRef(imageRef, imageDigest)
+		if err != nil {
+			logger.Error("Failed to prepare a canonical reference. Cannot pull image.", logger.Fields{
+				field.TaskID:      task.GetID(),
+				field.Container:   container.Name,
+				field.Image:       imageRef,
+				field.ImageDigest: imageDigest,
+				field.Error:       err,
 			})
 			return dockerapi.DockerContainerMetadata{
-				Error: dockerapi.CannotPullECRContainerError{
-					FromError: errors.New("engine ecr credentials: not found"),
+				Error: dockerapi.CannotPullContainerError{
+					FromError: fmt.Errorf(
+						"failed to prepare a canonical reference with image '%s' and digest '%s': %w",
+						imageRef, imageDigest, err),
 				},
 			}
 		}
-
-		iamCredentials := executionCredentials.GetIAMRoleCredentials()
-		container.SetRegistryAuthCredentials(iamCredentials)
-		// Clean up the ECR pull credentials after pulling
-		defer container.SetRegistryAuthCredentials(credentials.IAMRoleCredentials{})
+		imageRef = canonicalRef.String()
+		logger.Info("Prepared a canonical reference for image pull", logger.Fields{
+			field.TaskID:      task.GetID(),
+			field.Container:   container.Name,
+			field.Image:       container.Image,
+			field.ImageDigest: imageDigest,
+			field.ImageRef:    imageRef,
+		})
 	}
 
-	// Apply registry auth data from ASM if required
-	if container.ShouldPullWithASMAuth() {
-		if err := task.PopulateASMAuthData(container); err != nil {
-			logger.Error("Unable to acquire Docker registry credentials to pull image for container", logger.Fields{
-				field.TaskID:    task.GetID(),
-				field.Container: container.Name,
-				field.Image:     container.Image,
-				field.Error:     err,
-			})
-			return dockerapi.DockerContainerMetadata{
-				Error: dockerapi.CannotPullContainerAuthError{
-					FromError: errors.New("engine docker private registry credentials: not found"),
-				},
-			}
-		}
-		defer container.SetASMDockerAuthConfig(types.AuthConfig{})
-	}
-
-	metadata := engine.client.PullImage(engine.ctx, container.Image, container.RegistryAuthentication, engine.cfg.ImagePullTimeout)
+	metadata := engine.client.PullImage(engine.ctx, imageRef, container.RegistryAuthentication, engine.cfg.ImagePullTimeout)
 
 	// Don't add internal images(created by ecs-agent) into image manager state
 	if container.IsInternal() {
 		return metadata
 	}
 	pullSucceeded := metadata.Error == nil
+
+	if pullSucceeded && imageRef != container.Image {
+		// Resolved image manifest digest was used to pull the image.
+		// Tag the pulled image so that it can be found using the image reference in the task.
+		ctx, cancel := context.WithTimeout(engine.ctx, tagImageTimeout)
+		defer cancel()
+		err := engine.client.TagImage(ctx, imageRef, container.Image)
+		if err != nil {
+			logger.Error("Failed to tag image after pull", logger.Fields{
+				field.TaskID:    task.GetID(),
+				field.Container: container.Name,
+				field.Image:     container.Image,
+				field.ImageRef:  imageRef,
+				field.Error:     err,
+			})
+			if errors.Is(err, context.DeadlineExceeded) {
+				metadata.Error = &dockerapi.DockerTimeoutError{
+					Duration:   tagImageTimeout,
+					Transition: "pulled",
+				}
+			} else {
+				metadata.Error = &dockerapi.CannotPullContainerError{FromError: err}
+			}
+			pullSucceeded = false
+		} else {
+			logger.Info("Successfully tagged image", logger.Fields{
+				field.TaskID:    task.GetID(),
+				field.Container: container.Name,
+				field.Image:     container.Image,
+				field.ImageRef:  imageRef,
+			})
+		}
+	}
+
 	findCachedImage := false
 	if !pullSucceeded {
 		// Extend error message
@@ -1454,11 +1614,11 @@ func (engine *DockerTaskEngine) pullAndUpdateContainerReference(task *apitask.Ta
 		// 2. ImagePullBehavior is not set to always
 		// search the image in local cached images
 		if engine.cfg.DependentContainersPullUpfront.Enabled() && engine.cfg.ImagePullBehavior != config.ImagePullAlwaysBehavior {
-			if _, err := engine.client.InspectImage(container.Image); err != nil {
+			if _, err := engine.client.InspectImage(imageRef); err != nil {
 				logger.Error("Failed to find cached image for container", logger.Fields{
 					field.TaskID:    task.GetID(),
 					field.Container: container.Name,
-					field.Image:     container.Image,
+					field.Image:     imageRef,
 					field.Error:     err,
 				})
 				// Stop the task if the container is an essential container,
@@ -1472,7 +1632,7 @@ func (engine *DockerTaskEngine) pullAndUpdateContainerReference(task *apitask.Ta
 			logger.Info("Found cached image, use it directly for container", logger.Fields{
 				field.TaskID:    task.GetID(),
 				field.Container: container.Name,
-				field.Image:     container.Image,
+				field.Image:     imageRef,
 			})
 			findCachedImage = true
 		}
@@ -1487,6 +1647,53 @@ func (engine *DockerTaskEngine) pullAndUpdateContainerReference(task *apitask.Ta
 
 	engine.updateContainerReference(pullSucceeded, container, task.GetID())
 	return metadata
+}
+
+// Sets image registry auth credentials for the container.
+// Returns a cleanup function that may be called to clear the credentials from the container
+// when they are no longer needed. The cleanup function is `nil` if no credentials are needed
+// for the container.
+func (engine *DockerTaskEngine) setRegistryCredentials(
+	container *apicontainer.Container, task *apitask.Task,
+) (func(), apierrors.NamedError) {
+	var cleanup func()
+
+	// Set the credentials for pull from ECR if necessary
+	if container.ShouldPullWithExecutionRole() {
+		executionCredentials, ok := engine.credentialsManager.GetTaskCredentials(task.GetExecutionCredentialsID())
+		if !ok {
+			logger.Error("Unable to acquire ECR credentials to pull image for container", logger.Fields{
+				field.TaskID:    task.GetID(),
+				field.Container: container.Name,
+				field.Image:     container.Image,
+			})
+			return nil, dockerapi.CannotPullECRContainerError{
+				FromError: errors.New("engine ecr credentials: not found"),
+			}
+		}
+
+		iamCredentials := executionCredentials.GetIAMRoleCredentials()
+		container.SetRegistryAuthCredentials(iamCredentials)
+		cleanup = func() { container.SetRegistryAuthCredentials(credentials.IAMRoleCredentials{}) }
+	}
+
+	// Apply registry auth data from ASM if required
+	if container.ShouldPullWithASMAuth() {
+		if err := task.PopulateASMAuthData(container); err != nil {
+			logger.Error("Unable to acquire Docker registry credentials to pull image for container", logger.Fields{
+				field.TaskID:    task.GetID(),
+				field.Container: container.Name,
+				field.Image:     container.Image,
+				field.Error:     err,
+			})
+			return nil, dockerapi.CannotPullContainerAuthError{
+				FromError: errors.New("engine docker private registry credentials: not found"),
+			}
+		}
+		cleanup = func() { container.SetASMDockerAuthConfig(types.AuthConfig{}) }
+	}
+
+	return cleanup, nil
 }
 
 func (engine *DockerTaskEngine) updateContainerReference(pullSucceeded bool, container *apicontainer.Container, taskId string) {